rename : 폴더 이동 및 상수명 변경 #22
