--- src/widgets/lab-top-menu/index.ts
@@ -1,1 +1,0 @@
-export { LabMenuWrapper } from './ui/lab-menu-wrapper'
