+++ src/widgets/menu/index.ts
@@ -1,2 +1,3 @@
 export { PageGnb } from './ui/page-gnb'
 export { MobileBottomBar } from './ui/mobile-bottom-bar'
+export { LabMenuWrapper } from '../menu/ui/lab-menu-wrapper'
