+++ src/widgets/menu/ui/lab-menu-item.tsx
