+++ src/widgets/menu/ui/mobile-bottom-bar.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useRouter } from 'next/dist/client/components/navigation'
-import { SIDEMENU_ITEM } from '@/shared/constant/nav'
+import { GNB_ITEM } from '@/shared/constant/nav'
 import { PageGnbItem } from '@/widgets/menu/ui/page-gnb-item'
 
 export const MobileBottomBar = () => {
@@ -13,7 +13,7 @@ export const MobileBottomBar = () => {
       }
     >
       <ul className={'flex w-full items-center justify-around'}>
-        {SIDEMENU_ITEM.map(({ name, path, Icon }) => (
+        {GNB_ITEM.map(({ name, path, Icon }) => (
           <PageGnbItem device={'mobile'} itemName={name} path={path} onItemClick={() => router.push(path)}>
             <Icon />
           </PageGnbItem>
